FROMGIT: ASoC: SOF: update dai config for all dai with same name
Previously, dai config was updated only for one DAI with the same
name as the link name. But in the case of duplex pcm's, there will
be two DAI's with the same name and both of them need to be updated.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
(cherry picked from commit f8c0eb85da30403b4d9da56383280df0f2455ad4)
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>

Change-Id: Ieff3fe5418bd70ac957e6f8c089f49482df740ee